+++ src/Schema/Directives/MorphOneDirective.php
@@ -2,10 +2,7 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
-use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
-use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
-
-class MorphOneDirective extends RelationDirective implements FieldResolver, DefinedDirective
+class MorphOneDirective extends RelationDirective
 {
     public static function definition(): string
     {