+++ src/Schema/Directives/ScopeDirective.php
@@ -5,9 +5,8 @@ namespace Nuwave\Lighthouse\Schema\Directives;
 use BadMethodCallException;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
-use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
-class ScopeDirective extends BaseDirective implements ArgBuilderDirective, DefinedDirective
+class ScopeDirective extends BaseDirective implements ArgBuilderDirective
 {
     public static function definition(): string
     {
@@ -40,7 +39,7 @@ SDL;
             return $builder->{$scope}($value);
         } catch (BadMethodCallException $exception) {
             throw new DefinitionException(
-                $exception->getMessage()." in {$this->name()} directive on {$this->nodeName()} argument.",
+                $exception->getMessage()." in @{$this->name()} directive on {$this->nodeName()} argument.",
                 $exception->getCode(),
                 $exception->getPrevious()
             );